requirements.txt
requests>=2.31
Pillow>=10.0
PyMuPDF>=1.24.3
pymupdf4llm>=0.2.9
markitdown>=0.1.4
